backend/requirements.txt
fastapi
uvicorn
asyncpg
redis
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pyotp
cryptography
